+++ R/utils.R
@@ -70,7 +70,7 @@ list_sources <- function() {
       "1991, 2000, 2010, 2022",
       "1996-2024",
       "1996-2024",
-      "1992-2023",
+      "2008-2024",
       "1994-2023",
       "2001-2023",
       "2005-2023"
@@ -83,7 +83,7 @@ list_sources <- function() {
       "available",
       "available",
       "available",
-      "planned",
+      "available",
       "planned",
       "planned",
       "planned"